remove execute_hook, update hook_caller.__call__
