+++ napari/plugins/_tests/test_hook_specifications.py
@@ -54,6 +54,10 @@ def test_annotation_on_hook_specification(name, func):
     sig = inspect.signature(func)
     if sig.parameters:
         for param in sig.parameters.values():
+            for forbidden in ('_plugin', '_skip_impls', '_return_impl'):
+                assert (
+                    param.name != forbidden
+                ), f'Must not name hook_specification argument "{forbidden}"!'
             assert param.annotation is not param.empty, (
                 f"in hook specification '{name}', parameter '{param}' "
                 "has no type annotation"
