+++ napari/plugins/manager.py
@@ -3,15 +3,14 @@ import os
 import pkgutil
 import sys
 from logging import getLogger
-from typing import Generator, Optional, Tuple, Union, List, Dict, Any
+from typing import Dict, Generator, Optional, Tuple, Union
 
 import pluggy
-from pluggy.hooks import HookImpl
 
 from . import _builtins, hook_specifications
+from ._hook_callers import _HookCaller
 from .exceptions import (
     PluginError,
-    PluginCallError,
     PluginImportError,
     PluginRegistrationError,
     fetch_module_metadata,
@@ -23,206 +22,6 @@ if sys.version_info >= (3, 8):
     from importlib import metadata as importlib_metadata
 else:
     import importlib_metadata
-
-
-class _HookCaller(pluggy.hooks._HookCaller):
-    """Adding convenience methods to PluginManager.hook
-
-    In a pluggy plugin manager, the hook implementations registered for each
-    plugin are stored in ``_HookCaller`` objects that share the same name as
-    the corresponding hook specification; and each ``_HookCaller`` instance is
-    stored under the ``plugin_manager.hook`` namespace. For instance:
-    ``plugin_manager.hook.name_of_hook_specification``.
-    """
-
-    # just for type annotation.  These are the lists that store HookImpls
-    _wrappers: List[HookImpl]
-    _nonwrappers: List[HookImpl]
-
-    def get_hookimpl_for_plugin(self, plugin_name: str):
-        """Return hook implementation instance for ``plugin_name`` if found."""
-        try:
-            return next(
-                imp
-                for imp in self.get_hookimpls()
-                if imp.plugin_name == plugin_name
-            )
-        except StopIteration:
-            raise KeyError(
-                f"No implementation of {self.name} found "
-                f"for plugin {plugin_name}."
-            )
-
-    def index(self, value: Union[str, HookImpl]) -> int:
-        """Return index of plugin_name or a HookImpl in self._nonwrappers"""
-        if isinstance(value, HookImpl):
-            return self._nonwrappers.index(value)
-        elif isinstance(value, str):
-            plugin_names = [imp.plugin_name for imp in self._nonwrappers]
-            return plugin_names.index(value)
-        else:
-            raise TypeError(
-                "argument provided to index must either be the "
-                "(string) name of a plugin, or a HookImpl instance"
-            )
-
-    def bring_to_front(self, new_order: Union[List[str], List[HookImpl]]):
-        """Move items in ``new_order`` to the front of the call order.
-
-        By default, hook implementations are called in last-in-first-out order
-        of registration, and pluggy does not provide a built-in way to
-        rearrange the call order of hook implementations.
-
-        This function accepts a `_HookCaller` instance and the desired
-        ``new_order`` of the hook implementations (in the form of list of
-        plugin names, or a list of actual ``HookImpl`` instances) and reorders
-        the implementations in the hook caller accordingly.
-
-        NOTE: hook implementations are actually stored in *two* separate list
-        attributes in the hook caller: ``_HookCaller._wrappers`` and
-        ``_HookCaller._nonwrappers``, according to whether the corresponding
-        ``HookImpl`` instance was marked as a wrapper or not.  This method
-        *only* sorts _nonwrappers.
-        For more, see: https://pluggy.readthedocs.io/en/latest/#wrappers
-
-        Parameters
-        ----------
-        new_order :  list of str or list of ``HookImpl`` instances
-            The desired CALL ORDER of the hook implementations.  The list
-            does *not* need to include every hook implementation in
-            ``self.get_hookimpls()``, but those that are not included
-            will be left at the end of the call order.
-
-        Raises
-        ------
-        TypeError
-            If any item in ``new_order`` is neither a string (plugin_name) or a
-            ``HookImpl`` instance.
-        ValueError
-            If any item in ``new_order`` is neither the name of a plugin or a
-            ``HookImpl`` instance that is present in self._nonwrappers.
-        ValueError
-            If ``new_order`` argument has multiple entries for the same
-            implementation.
-
-        Examples
-        --------
-        Imagine you had a hook specification named ``print_plugin_name``, that
-        expected plugins to simply print their own name. An implementation
-        might look like:
-
-        >>> # hook implementation for ``plugin_1``
-        >>> @hook_implementation
-        ... def print_plugin_name():
-        ...     print("plugin_1")
-
-        If three different plugins provided hook implementations. An example
-        call for that hook might look like:
-
-        >>> plugin_manager.hook.print_plugin_name()
-        plugin_1
-        plugin_2
-        plugin_3
-
-        If you wanted to rearrange their call order, you could do this:
-
-        >>> new_order = ["plugin_2", "plugin_3", "plugin_1"]
-        >>> plugin_manager.hook.print_plugin_name.bring_to_front(new_order)
-        >>> plugin_manager.hook.print_plugin_name()
-        plugin_2
-        plugin_3
-        plugin_1
-
-        You can also just specify one or more item to move them to the front
-        of the call order:
-        >>> plugin_manager.hook.print_plugin_name.bring_to_front(["plugin_3"])
-        >>> plugin_manager.hook.print_plugin_name()
-        plugin_3
-        plugin_2
-        plugin_1
-        """
-        # make sure items in order are unique
-        if len(new_order) != len(set(new_order)):
-            raise ValueError("repeated item in order")
-
-        # make new lists for the rearranged _nonwrappers
-        # for details on the difference between wrappers and nonwrappers, see:
-        # https://pluggy.readthedocs.io/en/latest/#wrappers
-        _old_nonwrappers = self._nonwrappers.copy()
-        _new_nonwrappers: List[HookImpl] = []
-        indices = [self.index(elem) for elem in new_order]
-        for i in indices:
-            _new_nonwrappers.insert(0, _old_nonwrappers[i])
-
-        # remove items that have been pulled, leaving only items that
-        # were not specified in ``new_order`` argument
-        # do this rather than using .pop() above to avoid changing indices
-        for i in sorted(indices, reverse=True):
-            del _old_nonwrappers[i]
-
-        # if there are any hook_implementations left over, add them to the
-        # beginning of their respective lists
-        if _old_nonwrappers:
-            _new_nonwrappers = [x for x in _old_nonwrappers] + _new_nonwrappers
-
-        # update the _nonwrappers list with the reordered list
-        self._nonwrappers = _new_nonwrappers
-
-    def _set_plugin_enabled(self, plugin_name: str, enabled: bool):
-        """Enable or disable the hook implementation for a specific plugin.
-
-        Parameters
-        ----------
-        plugin_name : str
-            The name of a plugin implementing ``hook_spec``.
-        enabled : bool
-            Whether or not the implementation should be enabled.
-
-        Raises
-        ------
-        KeyError
-            If ``plugin_name`` has not provided a hook implementation for this
-            hook specification.
-        """
-        self.get_hookimpl_for_plugin(plugin_name).enabled = enabled
-
-    def enable_plugin(self, plugin_name: str):
-        """enable implementation for ``plugin_name``."""
-        self._set_plugin_enabled(plugin_name, True)
-
-    def disable_plugin(self, plugin_name: str):
-        """disable implementation for ``plugin_name``."""
-        self._set_plugin_enabled(plugin_name, False)
-
-    def call_plugin(self, plugin_name: str, *args, **kwargs):
-        implementation = self.get_hookimpl_for_plugin(plugin_name)
-        if implementation.hookwrapper:
-            raise TypeError("Hook wrappers can not be called directly")
-
-        # pluggy only allows calling hooks with keyword arguments
-        if args:
-            raise TypeError("hook calling supports only keyword arguments")
-        _args: List[Any] = []
-        try:
-            _args = [kwargs[argname] for argname in implementation.argnames]
-        except KeyError:
-            for argname in implementation.argnames:
-                if argname not in kwargs:
-                    raise pluggy.HookCallError(
-                        f"hook call must provide argument {argname}"
-                    )
-
-        try:
-            return implementation.function(*_args)
-        except Exception as exc:
-            raise PluginCallError(
-                implementation.plugin_name,
-                implementation.plugin.__name__,
-                msg=(
-                    f"Error calling plugin '{implementation.plugin_name}', "
-                    f"hook '{str(implementation.function.__name__)}'"
-                ),
-            ) from exc
 
 
 pluggy.manager._HookCaller = _HookCaller
